requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
attrs==24.2.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dnspython==2.7.0
flatbuffers==24.3.25
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.0
narwhals==1.14.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.5
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
pyee==12.1.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
streamlit==1.40.1
streamlit-webrtc==0.47.9
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
